Remove LRU cache
Using the LRU cache does not offer any benefit, as the settings are on initialized once at app startup:

https://github.com/stac-utils/titiler-pgstac/blob/780f4f8815f9349f72bb7805fad1306bb4412291/titiler/pgstac/main.py#L31
https://github.com/stac-utils/titiler-pgstac/blob/780f4f8815f9349f72bb7805fad1306bb4412291/titiler/pgstac/db.py#L12-L17
https://github.com/stac-utils/titiler-pgstac/blob/780f4f8815f9349f72bb7805fad1306bb4412291/titiler/pgstac/dependencies.py#L20

They additionally don't allow settings to be provided manually by keyword arguments.  If it is deemed preferable to have the LRU present, an alternative solution would be to alter the cached function to pass through kwargs to the settings class.